refactor: split edit-view templates
